refactor: decouple sys module of new napi-rs into a crate
