+++ .github/workflows/linux-musl.yaml
@@ -33,7 +33,7 @@ jobs:
 
       - name: Run tests
         run: |
-          docker run --rm -v $(pwd)/.cargo:/root/.cargo -v $(pwd):/napi-rs -w /napi-rs builder cargo test -p napi-rs --lib -- --nocapture
+          docker run --rm -v $(pwd)/.cargo:/root/.cargo -v $(pwd):/napi-rs -w /napi-rs builder cargo test -p napi-sys --lib -- --nocapture
 
       - name: Unit test
         run: |
