+++ .github/workflows/linux.yaml
@@ -58,7 +58,7 @@ jobs:
         timeout-minutes: 5
         with:
           command: test
-          args: -p napi-rs --lib -- --nocapture
+          args: -p napi-sys --lib -- --nocapture
 
       - name: Unit tests
         run: |
