+++ napi-derive-example/Cargo.toml
@@ -8,8 +8,8 @@ edition = "2018"
 crate-type = ["cdylib"]
 
 [dependencies]
-napi-rs = { path = "../napi" }
-napi-rs-derive = { path = "../napi-derive" }
+napi = { path = "../napi" }
+napi-derive = { path = "../napi-derive" }
 
 [build-dependencies]
 napi-build = { path = "../build" }
