+++ napi-derive/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
-name = "napi-rs-derive"
-version = "0.2.0"
+name = "napi-derive"
+version = "0.4.0"
 authors = ["LongYinan <lynweklm@gmail.com>"]
 edition = "2018"
 description = "N-API procedural macros"
