+++ napi/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
-name = "napi-rs"
-version = "0.3.1"
+name = "napi"
+version = "0.4.0"
 authors = ["Nathan Sobo <nathan@github.com>", "Yinan Long <lynweklm@gmail.com>"]
 license = "MIT"
 description = "N-API bindings"
@@ -12,6 +12,9 @@ edition = "2018"
 [features]
 libuv = ["futures"]
 tokio_rt = ["futures", "tokio", "once_cell"]
+
+[dependencies]
+napi-sys = { version = "0.4", path = "../sys" }
 
 [dependencies.futures]
 version = "0.3"
@@ -27,14 +30,5 @@ version = "1.4"
 optional = true
 
 
-[target.'cfg(windows)'.build-dependencies]
-flate2 = "1.0"
-reqwest = { version = "0.10", features = ["native-tls", "blocking"] }
-tar = "0.4"
-
 [build-dependencies]
-bindgen = "0.54"
-glob = "0.3"
 napi-build = { version = "0.1", path = "../build" }
-regex = "1.3"
-semver = "0.10"
