+++ sys/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "napi-sys"
+version = "0.4.0"
+authors = ["LongYinan <lynweklm@gmail.com>"]
+edition = "2018"
+readme = "README.md"
+repository = "https://github.com/Brooooooklyn/napi-rs"
+license = "MIT"
+keywords = ["NodeJS", "FFI", "NAPI", "n-api"]
+
+[target.'cfg(windows)'.build-dependencies]
+flate2 = "1.0"
+reqwest = { version = "0.10", features = ["native-tls", "blocking"] }
+tar = "0.4"
+
+[build-dependencies]
+bindgen = "0.54"
+glob = "0.3"
+regex = "1.3"
+semver = "0.10"
