+++ sys/README.md
@@ -1,0 +1,5 @@
+# napi-sys
+
+Low-level N-API bindings for Node.js addons written in Rust.
+
+See the [napi](https://github.com/napi-rs/napi-rs) for the high-level API.
