+++ test_module/Cargo.toml
@@ -9,8 +9,8 @@ crate-type = ["cdylib"]
 
 [dependencies]
 futures = "0.3"
-napi-rs = { path = "../napi", features = ["libuv", "tokio_rt"] }
-napi-rs-derive = { path = "../napi-derive" }
+napi = { path = "../napi", features = ["libuv", "tokio_rt"] }
+napi-derive = { path = "../napi-derive" }
 tokio = { version = "0.2", features = ["default", "fs"]}
 
 [build-dependencies]
